// aurora-eadi-front/src/services/faturamento/types/type_faturamentoDetalhado.ts
export type FaturamentoDetalhado = {
  id: number;
  cod_cli: string;
  n_fatura: string;
  cliente: string;
  endereco: string;
  bairro: string;
  cidade: string;
  uf: string;
  cep: string;
  cgc: string;
  inscr_esta: string;
  inscr_munic: string;
  rps: string;
  tp_nota: string;
  vl_extenso: string;
  valor_fatura: number;
  valor_servicos: number;
  dt_fatura: string;
  dt_vencimento: string;
  iss_cobrar: string;
  iss_valor: number;
  iss_percentual: number;
  ii_valor: number;
  observacao: string;
  modalidade: string;
  despachante: string;
  valor_cif: number;
  n_da: string;
  n_di: string;
  n_lote: string;
  n_conhecimento: string;
  n_documento: string;
  tx_dolar: number;
  dt_entrada: string;
  nr_periodo_i: number;
  nr_periodo_f: number;
  dt_periodo_f: string;
  qt_volumes: number;
  peso_bruto: number;
  m3: number;
  tributacao_msg: string;
  quantidade: number;
  servico_id: string;
  servico: string;
  valor: number;
  modalidade_txt: string;
  cliente_rps: number;
};